Move `*Result` objects in `API\Model` to `API\Model\Results`
